5/2/2019 Changes *SEE DESCRIPTION*
This is a small change to some code inside of the userMenu, fileHandling, and budget files. The findCorrect file is new though. This class does not any new functionality to the program but does clean up some code that is used multiple times inside of the other classes. That is now all in one place making the already crazy user menu class a little cleaner. In the next updates I plan on gradually cleaning up the userMenu class and breaking it up into multiple files to make things more readable to myself and anyone else working on the code of the program